pyproject.toml
[project]
name = "ingest-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-storage>=3.1.0",
    "jinja2>=3.1.6",
    "pandas>=2.2.3",
    "pandas-gbq>=0.29.1",
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.2",
    "ruff>=0.11.6",
    "sqlalchemy>=2.0.40",
]
